Fix: 수정 페이지에서도 사용하기 위해 컴포넌트 분리
